requirements.txt
numpy==1.21.4
pandas==1.5.1
Pillow==9.3.0
streamlit==1.15.0
